requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
amqp==1.4.9               # via kombu
analytics-python==1.2.9   # via -c requirements/constraints.txt, -r requirements/base.in
anyjson==0.3.3            # via kombu
billiard==3.3.0.23        # via celery
celery==3.1.26.post2      # via -c requirements/constraints.txt, -r requirements/base.in, django-user-tasks
certifi==2020.6.20        # via -r requirements/github.txt, requests
cffi==1.14.1              # via cryptography
chardet==3.0.4            # via -r requirements/github.txt, requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
cryptography==3.0         # via pyjwt
defusedxml==0.6.0         # via -r requirements/github.txt, python3-openid, social-auth-core
django-cors-headers==3.4.0  # via -r requirements/base.in
django-extensions==3.0.4  # via -r requirements/base.in
django-guardian==2.3.0    # via -r requirements/base.in
django-model-utils==4.0.0  # via -r requirements/base.in, django-user-tasks
django-mysql==3.8.1       # via -r requirements/base.in
django-simple-history==2.11.0  # via -r requirements/base.in
django-storages==1.8      # via -c requirements/constraints.txt, -r requirements/base.in
django-user-tasks==1.1.0  # via -r requirements/base.in
django-waffle==1.0.0      # via -r requirements/base.in, edx-django-utils, edx-drf-extensions
django==2.2.15            # via -c requirements/constraints.txt, -r requirements/base.in, django-cors-headers, django-guardian, django-model-utils, django-mysql, django-storages, django-user-tasks, djangorestframework, drf-jwt, drf-yasg, edx-api-doc-tools, edx-auth-backends, edx-django-release-util, edx-django-utils, edx-drf-extensions, rest-condition
djangorestframework==3.11.1  # via -r requirements/base.in, django-user-tasks, drf-jwt, drf-yasg, edx-api-doc-tools, edx-drf-extensions, rest-condition
drf-jwt==1.16.2           # via edx-drf-extensions
drf-yasg==1.17.1          # via edx-api-doc-tools
edx-api-doc-tools==1.3.1  # via -r requirements/base.in
edx-auth-backends==3.1.0  # via -r requirements/base.in
edx-django-release-util==0.4.4  # via -r requirements/base.in
edx-django-utils==3.6.0   # via edx-drf-extensions
edx-drf-extensions==6.1.1  # via -r requirements/base.in
edx-opaque-keys==2.1.1    # via edx-drf-extensions
edx-rest-api-client==1.9.2  # via -c requirements/constraints.txt, -r requirements/base.in
future==0.18.2            # via pyjwkest
idna==2.8                 # via -c requirements/constraints.txt, -r requirements/github.txt, requests
inflection==0.5.0         # via drf-yasg
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
kombu==3.0.37             # via celery
markupsafe==1.1.1         # via jinja2
newrelic==5.14.1.144      # via edx-django-utils
oauthlib==3.1.0           # via -r requirements/github.txt, requests-oauthlib, social-auth-core
packaging==20.4           # via drf-yasg
pbr==5.4.5                # via stevedore
psutil==1.2.1             # via edx-django-utils
pycparser==2.20           # via cffi
pycryptodomex==3.9.8      # via pyjwkest
pyjwkest==1.4.2           # via edx-drf-extensions
pyjwt[crypto]==1.7.1      # via -r requirements/github.txt, drf-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pymongo==3.11.0           # via edx-opaque-keys
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via analytics-python, edx-drf-extensions
python3-openid==3.2.0     # via -r requirements/github.txt, social-auth-core
pytz==2020.1              # via -r requirements/base.in, celery, django
pyyaml==5.1               # via -c requirements/constraints.txt, edx-django-release-util
redis==2.8                # via -c requirements/constraints.txt, -r requirements/base.in
requests-oauthlib==1.3.0  # via -r requirements/github.txt, social-auth-core
requests==2.24.0          # via -r requirements/github.txt, analytics-python, coreapi, edx-drf-extensions, edx-rest-api-client, pyjwkest, requests-oauthlib, slumber, social-auth-core
rest-condition==1.0.3     # via edx-drf-extensions
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via drf-yasg
semantic-version==2.8.5   # via edx-drf-extensions
six==1.15.0               # via -r requirements/github.txt, analytics-python, cryptography, django-simple-history, drf-yasg, edx-auth-backends, edx-django-release-util, edx-drf-extensions, edx-opaque-keys, packaging, pyjwkest, python-dateutil, social-auth-app-django, social-auth-core
slumber==0.7.1            # via edx-rest-api-client
git+https://github.com/python-social-auth/social-app-django.git@c00d23c2b45c3317bd35b15ad1b959338689cef8#egg=social-auth-app-django  # via -r requirements/github.txt, edx-auth-backends
social-auth-core==3.2.0   # via -c requirements/constraints.txt, -r requirements/github.txt, edx-auth-backends, social-auth-app-django
sqlparse==0.3.1           # via django
stevedore==3.2.0          # via edx-opaque-keys
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.10          # via -r requirements/github.txt, requests
